Add python-nostr submodule and refactor twitter auth
