+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "python-nostr"]
+	path = python-nostr
+	url = https://github.com/jeffthibault/python-nostr.git
